issue-2744 updated changes with new description
