+++ test/load/allocation/configure-agones.sh
