--- test/load/allocation/grpc/fleet.yaml
@@ -1,47 +1,0 @@
-# Copyright 2020 Google LLC All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: "agones.dev/v1"
-kind: Fleet
-metadata:
-  name: load-test-fleet
-spec:
-  # the number of GameServers to keep Ready or Allocated in this Fleet
-  replicas: 4000
-  template:
-    metadata:
-      labels:
-        foo: bar
-    # GameServer specification
-    spec:
-        ports:
-        - containerPort: 7654
-          name: default
-        # The GameServer's Pod template
-        template:
-            spec:
-                containers:
-                - args:
-                  # We setup the simple-game-server server to shutdown 10 mins after allocation
-                  - -automaticShutdownDelaySec=600
-                  image: gcr.io/agones-images/simple-game-server:0.14
-                  name: simple-game-server
-                  resources:
-                    limits:
-                      cpu: 20m
-                      memory: 32Mi
-                    requests:
-                      cpu: 20m
-                      memory: 32Mi
-
