--- test/load/allocation/grpc/variable.txt
@@ -1,48 +1,0 @@
-# Copyright 2022 Google LLC All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-### Varying allocations
-#Duration,Number_of_clients/allocations
-# run for 20 mins with 10 clients
-10m,10
-10m,10
-# run for 30 mins with 30 clients
-10m,30
-10m,30
-10m,30
-# run for 30 mins with 20 clients
-10m,20
-10m,20
-10m,20
-# run for 30 mins with 30 clients
-10m,30
-10m,30
-10m,30
-# run for 30 mins with 20 clients
-10m,20
-10m,20
-10m,20
-# run for 30 mins with 5 clients
-10m,5
-10m,5
-10m,5
-# run for 40 mins with 30 clients
-10m,30
-10m,30
-10m,30
-10m,30
-# run for 30 mins with 15 clients
-10m,15
-10m,15
-10m,15
